Added LogProcessor for [Log] tags
Added as a built-in feature LogProcessor compatible for Swiz v1. Rather than use a custom processor, this processor will allow developers to easily inject a common/shared ILogger instance where [Log] is used. With [Log] Developers do not need to worry about creating a logger instance; it is magically injected and available with syntax like:
  
   log.debug('some message here.');
 
@TODO:  Still need to support [Log("targetServices")] to support an internal lookup of a loggerTarget by specifying a custom id.