Synch w/ Upstream 10/09/2010 Changes
Made tear down more efficient by using the view to Swiz instance association that is now created during set up
Added ISetUpValidator and ITearDownValidator interfaces to allow per-type opt out of the automatic set up and tear down process
Event Chaining API changes.
  Added IAsyncChainStep interface.
  Added IChainableEvent interface and base ChainableEvent implementation.
  Added missing copyright / license header to EventTypeExpression and IAutonomousChainStep.
  Added asdoc comments and comment headers to BaseCompositeChain, CommandChain.
  Added IAsyncChainStep / IChainableEvent handling to Mediator.
  Modified EventChain to interact with the new EventChainStep implementation.
  Rewrote EventChainStep to implement IAutonomousChainStep, IAsyncChainStep and extend BaseChainStep rather than Event.
  EventChainStep can now be used independently of EventChain in conjunction with any CompositeChain.
  EventChainStep can now optionally specify its own dispatcher, independent of an EventChain.
  EventChainStep can now dispatch an IChainableEvent and await the results of multiple pending asynchronous operations in [Mediate]'ed methods before completing.
  FunctionChainStep now captures and provides the function return value as a property.
  Renamed IChainableEvent/ChainableEvent to IAsynchronousEvent/AsynchronousEvent for clarity.
  Modified Mediator to support mmethod return values of AsyncToken or IAsyncOperation when mediating an IAsynchronousEvent.
  Added IAsynchronousOperation interface and AbstractAsynchronousOperation base class.
  Added concrete implementations of IAsynchronousOperation: AsyncTokenOperation (for adapting AsyncToken instances), AsynchronousChainOperation (for adapting IChain instances) and AsynchronousIOOperation (for adapting FileReference, Loader, etc. instances).
  Added AbstractAsynchronousDispatcherOperation to simplify creating IAsynchronousOperations adapters that are event based.
  Modified IAsyncChainStep (and EventChainStep) to use IAsynchronousOperation instead of AsyncToken.
  Added missing [Event] metadata and 'extends IEventDispatcher' to IChain.
  Removed unnecessary imports.
  Fixed minor formatting error in license comment block for newly added classes.
Updated URLRequestHelper to match ServiceHelper api, and return the original request. Also updated the AbstractController methods to match both methods from ServiceHelper and URLRequestHelper.
Ignore classes whose names contain double underscores. When embedding an image in MXML, the compiler represents it as a class whose name matches the containing component, plus some text that includes __. This is the flag we're using to prevent the attempted processing of embedded images.
Wired views are processed as beans, but are not actually added to the beans list. Therefore, getBeanForSource() cannot find a bean for them and null is passed to tearDownBean(). For now we will just create a fresh Bean instance as it serves the purpose we need.
If a user has not declared a bean for one of Swiz's built in helper classes, Swiz will now create one and supply it automatically. Code only runs if a typed injection was requested and not found.
Fixed issue with using addBeanProvider() when the beans contained depend on each other.