Updated readme and Fix to InjectorProcessor to support source propertyChains
InjectorProcessor would not always support propertyChains in metadata. If the property chain has elements that are NULL
   e.g. [Inject(source="employeeModel.selectedEmployee.address")] where selectedEmployee is null at setup,

then the setUpMetadataTag() throws an exception and would not configure data bindings properly. Now a warning is issued to the user and the the ChangeWatcher still attempts to establish databinding for property chains.